lessen main jsx and change directory of components
