+++ app/javascript/packs/components/chat/ChatDisplay.jsx
